knowledge-network-system/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.14.0
python-pptx==1.0.2
PyMuPDF==1.24.14
pdfplumber==0.11.4
networkx==3.4.2
numpy==2.1.3
httpx==0.28.1
orjson==3.10.12
pytest==8.3.4
